+++ service/post-service/src/main/java/com/example/post/mapper/CommentMapper.java
@@ -19,15 +19,15 @@ public interface CommentMapper extends BaseMapper<Comment> {
     /**
      * 分页查询帖子的评论
      */
-    IPage<Comment> getCommentsByPostId(Page<Comment> page, @Param("postId") Integer postId);
+    IPage<Comment> getCommentsByPostId(Page<Comment> page, @Param("postId") Long postId);
 
     /**
      * 查询用户发布的评论
      */
-    IPage<Comment> getUserComments(Page<Comment> page, @Param("userId") Integer userId);
+    IPage<Comment> getUserComments(Page<Comment> page, @Param("userId") Long userId);
 
     /**
      * 删除评论（软删除）
      */
-    int deleteComment(@Param("commentId") Integer commentId);
+    int deleteComment(@Param("commentId") Long commentId);
 }
