+++ service/post-service/src/main/java/com/example/post/mapper/MessageMapper.java
@@ -22,31 +22,31 @@ public interface MessageMapper extends BaseMapper<Message> {
      * 分页查询用户的私信
      */
     IPage<Message> getUserMessages(Page<Message> page, 
-                                 @Param("userId") Integer userId, 
-                                 @Param("otherUserId") Integer otherUserId);
+                                 @Param("userId") Long userId, 
+                                 @Param("otherUserId") Long otherUserId);
 
     /**
      * 获取用户的未读消息数量
      */
-    int getUnreadMessageCount(@Param("userId") Integer userId);
+    int getUnreadMessageCount(@Param("userId") Long userId);
 
     /**
      * 标记消息为已读
      */
-    int markMessageAsRead(@Param("messageId") Integer messageId);
+    int markMessageAsRead(@Param("messageId") Long messageId);
 
     /**
      * 标记与某用户的所有消息为已读
      */
-    int markAllMessagesAsRead(@Param("userId") Integer userId, @Param("otherUserId") Integer otherUserId);
+    int markAllMessagesAsRead(@Param("userId") Long userId, @Param("otherUserId") Long otherUserId);
 
     /**
      * 删除私信（软删除）
      */
-    int deleteMessage(@Param("messageId") Integer messageId);
+    int deleteMessage(@Param("messageId") Long messageId);
 
     /**
      * 获取用户的对话列表
      */
-    List<Message> getConversationList(@Param("userId") Integer userId);
+    List<Message> getConversationList(@Param("userId") Long userId);
 }
